+++ CHANGELOG.md
@@ -1,8 +1,16 @@
 # amanobot changelog
 
+## 1.4.0 (2019-07-29)
+- Bot API 4.4
+- Added the `setChatDescription` method.
+- Added the fields `can_send_polls`, `can_change_info`, `can_invite_users`, `can_pin_messages` and `permissions` to the `restrictChatMember` method.
+- Added the `ChatPermissions` object.
+- Removed the `all_members_are_administrators` field from the `Chat` object.
+- Added the `is_animated` field to the `Sticker` and `StickerSet` objects.
+
 ## 1.3.0 (2019-06-01)
 - Bot API 4.3
-- Added missing Passport related methods and objects 
+- Added missing Passport related methods and objects
 - Added the `LoginUrl` object
 - Added the `login_url` field to the `InlineKeyboardButton` object